added swarm plots for individual compound classes
the script now includes a section in which the swarm plots are done for each compound class, not just on the entire set of core/satellite species. 
These plots (screenshot uploaded as merge2 or merge1_compoundclass) visualize potential differences within particular compound classes between C and S (i.e. a refined view after looking at relative composition of Core vs. Satellite. 
As before, this is done for different methods of C/S designation, and water and porewater samples are split.